stacking_sats/pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stacking_sats"
version = "0.1.0"
description = "A flexible Bitcoin investment strategy backtesting library"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Matt Faltyn", email = "your.email@example.com"}
]
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Office/Business :: Financial :: Investment",
    "License :: OSI Approved :: MIT License",
]
dependencies = [
    "pandas>=1.3.0",
    "numpy>=1.20.0",
    "matplotlib>=3.4.0,<4.0.0",
    "pyarrow>=7.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "isort>=5.10.0",
    "mypy>=1.0.0",
    "pytest-cov>=4.0.0",
]
data = [
    "cryptocompare>=0.7.0",
    "yfinance>=0.2.0",
]

[project.urls]
Homepage = "https://github.com/mattfaltyn/stacking_sats"
Issues = "https://github.com/mattfaltyn/stacking_sats/issues"

[project.scripts]
stacksats = "stacking_sats.cli.main:main"

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310", "py311"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
strict_optional = true

[[tool.mypy.overrides]]
module = [
    "matplotlib.*",
    "numpy.*",
    "pandas.*",
    "pytest.*",
    "cryptocompare.*",
    "yfinance.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "--cov=stacking_sats --cov-report=term-missing --cov-report=xml"
testpaths = ["tests"] 